# [PERF] Implement Efficient Caching Policy for Static Assets