--- a/index.html
+++ b/index.html
@@ -29,13 +29,13 @@
   
   <!-- Stylesheets -->
   <script src="https://unpkg.com/@tailwindcss/browser@4"></script>
-  <link rel="stylesheet" href="/css/custom.css">
+  <link rel="stylesheet" href="css/custom.css?v&#x3D;d5410f46">
   
   <!-- Performance Optimization Script -->
-  <script src="/js/performance.js" defer></script>
+  <script src="js/performance.js?v&#x3D;96aaf719" defer></script>
   
   <!-- Advanced Analytics Script -->
-  <script src="/js/analytics.js" defer></script>
+  <script src="js/analytics.js?v&#x3D;a2e38f08" defer></script>
   
   <!-- Math Support -->
   
@@ -117,7 +117,7 @@
           <!-- Date Column -->
           <div class="flex-shrink-0 w-24 pt-1">
             <time class="text-sm font-serif" style="color: var(--warm-text-muted);">
-              2025-06-11
+              Jun 11
             </time>
           </div>
           
@@ -161,7 +161,7 @@
           <!-- Date Column -->
           <div class="flex-shrink-0 w-24 pt-1">
             <time class="text-sm font-serif" style="color: var(--warm-text-muted);">
-              2025-06-10
+              Jun 10
             </time>
           </div>
           
@@ -204,7 +204,7 @@
           <!-- Date Column -->
           <div class="flex-shrink-0 w-24 pt-1">
             <time class="text-sm font-serif" style="color: var(--warm-text-muted);">
-              2025-05-30
+              May 30
             </time>
           </div>