--- a/tests/build-site.test.js
+++ b/tests/build-site.test.js
@@ -514,6 +514,7 @@ describe('SiteBuilder', () => {
         scheduledPosts: [{ id: '2' }]
       });
       siteBuilder.copyStaticAssetsIncremental = jest.fn().mockResolvedValue();
+      siteBuilder.generateAssetVersions = jest.fn().mockResolvedValue();
       siteBuilder.generateHomePage = jest.fn().mockResolvedValue();
       siteBuilder.generateBlogPages = jest.fn().mockResolvedValue();
       siteBuilder.generatePostPages = jest.fn().mockResolvedValue();
@@ -528,6 +529,7 @@ describe('SiteBuilder', () => {
       expect(siteBuilder.loadTemplates).toHaveBeenCalled();
       expect(siteBuilder.loadContent).toHaveBeenCalled();
       expect(siteBuilder.copyStaticAssetsIncremental).toHaveBeenCalled();
+      expect(siteBuilder.generateAssetVersions).toHaveBeenCalled();
       expect(siteBuilder.generateHomePage).toHaveBeenCalled();
       expect(siteBuilder.generateBlogPages).toHaveBeenCalled();
       expect(siteBuilder.generatePostPages).toHaveBeenCalled();