--- a/vercel.json
+++ b/vercel.json
@@ -66,7 +66,16 @@
       ]
     },
     {
-      "source": "/(.*\\.(ico|png|jpg|jpeg|gif|svg|webp|woff|woff2|eot|ttf|otf))",
+      "source": "/*.(ico|png|jpg|jpeg|gif|svg|webp)",
+      "headers": [
+        {
+          "key": "Cache-Control",
+          "value": "public, max-age=31536000, immutable"
+        }
+      ]
+    },
+    {
+      "source": "/*.(woff|woff2|eot|ttf|otf)",
       "headers": [
         {
           "key": "Cache-Control",