--- a/vercel.json
+++ b/vercel.json
@@ -66,7 +66,7 @@
       ]
     },
     {
-      "source": "/*.(ico|png|jpg|jpeg|gif|svg|webp)",
+      "source": "/static/(.*)",
       "headers": [
         {
           "key": "Cache-Control",
@@ -75,7 +75,7 @@
       ]
     },
     {
-      "source": "/*.(woff|woff2|eot|ttf|otf)",
+      "source": "/src/static/(.*)",
       "headers": [
         {
           "key": "Cache-Control",